refactor: application 레이어 계층 분리
